Extract spec Helpers module to a support file
